+++ test/libsolidity/syntaxTests/parsing/address_payable_library.sol
@@ -1,0 +1,5 @@
+library L {
+}
+contract C {
+    using L for address payable;
+}