+++ test/libsolidity/syntaxTests/parsing/address_payable_state_variable.sol
@@ -1,0 +1,8 @@
+contract C {
+    address payable a;
+    address payable public b;
+    address payable[] c;
+    address payable[] public d;
+    mapping(uint => address payable) e;
+    mapping(uint => address payable[]) f;
+}