+++ test/libsolidity/syntaxTests/parsing/address_payable_type_expression.sol
@@ -1,0 +1,7 @@
+contract C {
+    function f() public pure {
+        address payable;
+    }
+}
+// ----
+// ParserError: (67-68): Expected identifier but got ';'