+++ test/libsolidity/syntaxTests/parsing/mapping_from_address_payable.sol
@@ -1,0 +1,5 @@
+contract C {
+    mapping(address payable => uint) m;
+}
+// ----
+// ParserError: (33-40): Expected '=>' but got 'payable'